+++ mod-source-record-manager-server/src/main/java/org/folio/services/mappers/RecordToInstanceMapper.java
@@ -2,6 +2,7 @@ package org.folio.services.mappers;
 
 import io.vertx.core.json.JsonObject;
 import org.folio.rest.jaxrs.model.Instance;
+import org.folio.services.mappers.processor.parameters.MappingParameters;
 import org.folio.services.parsers.RecordFormat;
 
 /**
@@ -16,7 +17,7 @@ public interface RecordToInstanceMapper {
    * @return - Wrapper for parsed record in json format.
    * Can contains errors descriptions if parsing was failed
    */
-  Instance mapRecord(JsonObject parsedRecord);
+  Instance mapRecord(JsonObject parsedRecord, MappingParameters mappingParameters);
 
   /**
    * Provides access to the MapperFormat